fix(widgetbook): split trade screen catalog to satisfy 1000-line lint
Extract trade screen Widgetbook stories from catalog_data_screens.dart
into catalog_data_screens_trade.dart so repo.dart_file_non_comment_line_size
passes after #4226 additions.
